# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ContextForge API
  description: |
    ContextForge is an enterprise-grade, high-performance RAG and context-engine designed
    specifically for AI coding agents and developers.
    All data is strictly isolated by project and secured with single-user ownership in v1.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: https://api.contextforge.example.com/api/v1
    description: Production Server
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /auth/pat:
    post:
      summary: Authenticate via GitHub PAT Fallback
      description: Validates and saves a GitHub Personal Access Token for local development mode.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pat]
              properties:
                pat:
                  type: string
                  description: GitHub Personal Access Token with repo read scope.
      responses:
        '200':
          description: PAT validated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/401Problem'
  /auth/auto:
    get:
      summary: Auto-Authenticate Development Session
      description: Provisions a JWT session using the server's configured GITHUB_PAT without manual copy-paste.
      security: []
      responses:
        '200':
          description: Developer session provisioned or status returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401Problem'
  /auth/me:
    get:
      summary: Get Current User Profile
      description: Returns profile information and available installations for the authenticated user.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401Problem'
...
  /projects:
    get:
      summary: List User Projects
      description: Returns all projects owned by the authenticated user.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
                  total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
    post:
      summary: Create Project
      description: Creates a new isolated project container.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/400Problem'
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    get:
      summary: Get Project Details
      description: Fetches project metadata and statistics.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/404Problem'
    put:
      summary: Update Project
      description: Updates project configuration, name, description, or default provider settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete Project
      description: Cascading deletion of a project, all its sources, documents, and vector embeddings.
      responses:
        '204':
          description: Project deleted successfully.

  # ---------------------------------------------------------------------------
  # Sources
  # ---------------------------------------------------------------------------
  /projects/{id}/sources:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    get:
      summary: List Project Sources
      description: Lists all repositories and sources attached to the project.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Source' }
    post:
      summary: Connect Source Repository
      description: Connects a GitHub repository to the project and triggers initial ingestion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '202':
          description: Source connected and ingestion queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceIngestionQueued'
  /projects/{id}/sources/{source_id}/sync:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
      - name: source_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Trigger Source Synchronization
      description: Enqueues an asynchronous incremental or full synchronization job for the source.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force_full:
                  type: boolean
                  default: false
                  description: Force full re-chunking and re-embedding instead of incremental sync.
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'

  # ---------------------------------------------------------------------------
  # Database Sources
  # ---------------------------------------------------------------------------
...
  /projects/{id}/sources/database:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    get:
      summary: List Database Sources
      description: Lists all external database knowledge sources attached to the project.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DatabaseSource' }
    post:
      summary: Create Database Source
      description: Registers an external database knowledge source with encrypted credentials and tests connection before saving.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatabaseSourceRequest'
      responses:
        '201':
          description: Database source registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseSource'
...
  /projects/{id}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    get:
      summary: List Project Documents
      description: Query ingested documents with metadata filtering.
      parameters:
        - name: file_path
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Document' }
                  total: { type: integer }
  /projects/{id}/documents/{document_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
      - name: document_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get Document and Chunks
      description: Retrieves document content and its generated semantic chunks.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'

  # ---------------------------------------------------------------------------
  # Jobs & Queue
  # ---------------------------------------------------------------------------
  /jobs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get Job Status
      description: Polls the progress and execution state of an asynchronous ingestion or sync job.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'

  # ---------------------------------------------------------------------------
  # Chat & RAG Streaming
  # ---------------------------------------------------------------------------
  /projects/{id}/chat/completions:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    post:
      summary: RAG Query (Synchronous JSON)
      description: Executes vector similarity retrieval, constructs prompt with citations, and returns completed response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
  /projects/{id}/chat/completions/stream:
    parameters:
      - $ref: '#/components/parameters/ProjectIDPath'
    post:
      summary: RAG Query (Server-Sent Events Stream)
      description: Streams tokens and structured citations via text/event-stream.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: SSE stream of tokens and citations.
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: citation
                  data: {"file_path":"pkg/auth/jwt.go","start_line":42,"end_line":68,"similarity":0.89}

                  event: message
                  data: {"delta":"The authentication system uses AES-256-GCM..."}

                  event: done
                  data: {"total_tokens":420,"duration_ms":1450}

  # ---------------------------------------------------------------------------
  # GitHub Webhook Receiver
  # ---------------------------------------------------------------------------
  /github/webhooks:
    post:
      summary: GitHub Webhook Ingress
      description: Ingests push, pull_request, and installation events signed by GitHub HMAC-SHA256 secret.
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Event
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '202':
          description: Event accepted and queued for processing.
        '401':
          description: Invalid HMAC signature.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionCookie:
      type: apiKey
      in: cookie
      name: cf_session
  parameters:
    ProjectIDPath:
      name: id
      in: path
      required: true
      description: Unique Project UUID
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        default: 20
  responses:
    400Problem:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    401Problem:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    404Problem:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://contextforge.dev/errors/validation" }
        title: { type: string, example: "Validation Error" }
        status: { type: integer, example: 400 }
        detail: { type: string, example: "Project name must not be empty" }
        instance: { type: string, example: "/api/v1/projects" }
    User:
      type: object
      required: [id, github_login, email]
      properties:
        id: { type: string, format: uuid }
        github_login: { type: string, example: "octocat" }
        email: { type: string, format: email, example: "octocat@github.com" }
        name: { type: string, example: "The Octocat" }
        avatar_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string, description: "JWT session token" }
        user: { $ref: '#/components/schemas/User' }
    Project:
      type: object
      required: [id, name, owner_user_id, embedding_dimension, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "Backend Core Service" }
        description: { type: string, example: "Go RAG API and worker infrastructure" }
        owner_user_id: { type: string, format: uuid }
        embedding_provider: { type: string, example: "ollama" }
        embedding_model: { type: string, example: "nomic-embed-text" }
        embedding_dimension: { type: integer, example: 768 }
        llm_provider: { type: string, example: "cli_opencode" }
        total_documents: { type: integer, example: 142 }
        total_chunks: { type: integer, example: 1890 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        embedding_provider: { type: string, default: "ollama" }
        embedding_model: { type: string, default: "nomic-embed-text" }
        embedding_dimension: { type: integer, default: 768 }
        llm_provider: { type: string, default: "cli_opencode" }
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        llm_provider: { type: string }
    Source:
      type: object
      required: [id, project_id, name, repo_owner, repo_name, branch, sync_status]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string, example: "ContextForge Core" }
        repo_owner: { type: string, example: "your-org" }
        repo_name: { type: string, example: "contextforge" }
        branch: { type: string, example: "main" }
        last_commit_hash: { type: string, example: "a1b2c3d4e5f6..." }
        sync_status:
          type: string
          enum: [idle, queued, syncing, synced, failed]
        last_synced_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    CreateSourceRequest:
      type: object
      required: [repo_owner, repo_name]
      properties:
        repo_owner: { type: string }
        repo_name: { type: string }
        branch: { type: string, default: "main" }
    SourceIngestionQueued:
      type: object
      required: [source, job_id]
      properties:
        source: { $ref: '#/components/schemas/Source' }
        job_id: { type: string, format: uuid }
    Document:
      type: object
      required: [id, project_id, source_id, file_path, content_hash]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        file_path: { type: string, example: "internal/service/rag.go" }
        language: { type: string, example: "go" }
        content_hash: { type: string, example: "sha256:abcd..." }
        total_chunks: { type: integer, example: 5 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/Document'
        - type: object
          properties:
            chunks:
              type: array
              items: { $ref: '#/components/schemas/DocumentChunk' }
    DocumentChunk:
      type: object
      required: [id, chunk_index, start_line, end_line, content]
      properties:
        id: { type: string, format: uuid }
        chunk_index: { type: integer, example: 0 }
        start_line: { type: integer, example: 1 }
        end_line: { type: integer, example: 45 }
        content: { type: string }
        token_count: { type: integer, example: 380 }
    IngestionJob:
      type: object
      required: [id, project_id, source_id, status, progress_percent]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress_percent: { type: integer, minimum: 0, maximum: 100 }
        processed_files: { type: integer }
        total_files: { type: integer }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
    ChatCompletionRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "How does the token encryption work?" }
        top_k: { type: integer, default: 5 }
        similarity_threshold: { type: number, default: 0.7 }
        file_filters:
          type: array
          items: { type: string }
          example: ["internal/auth/*"]
        temperature: { type: number, default: 0.2 }
    Citation:
      type: object
      required: [source_id, file_path, start_line, end_line, similarity, snippet]
      properties:
        source_id: { type: string, format: uuid }
        file_path: { type: string, example: "internal/crypto/encrypt.go" }
        start_line: { type: integer, example: 12 }
        end_line: { type: integer, example: 35 }
        similarity: { type: number, example: 0.884 }
        snippet: { type: string }
    ChatCompletionResponse:
      type: object
      required: [answer, citations, tokens_used, duration_ms]
      properties:
        answer: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        tokens_used: { type: integer, example: 512 }
        duration_ms: { type: integer, example: 1250 }
...
    DatabaseSource:
      type: object
      required: [id, source_id, project_id, name, database_type, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string, example: "Production PostgreSQL" }
        database_type:
          type: string
          enum: [postgres, mysql, mariadb, sqlite, sqlserver, cockroachdb]
        host: { type: string, example: "db.example.com" }
        port: { type: integer, example: 5432 }
        database_name: { type: string, example: "app_db" }
        username: { type: string, example: "readonly_user" }
        configuration: { type: object }
        status:
          type: string
          enum: [created, configured, syncing, ready, failed]
        last_error: { type: string }
        last_synced_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateDatabaseSourceRequest:
      type: object
      required: [name, database_type, connection_url]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        database_type:
          type: string
          enum: [postgres, mysql, mariadb, sqlite, sqlserver, cockroachdb]
        connection_url: { type: string }
        configuration: { type: object }
...